app/DrinkConsumption/views.py
# from django.shortcuts import render

from .models import *
from .serializers import *

from rest_framework import viewsets
from rest_framework import permissions


class ContainerViewSet(viewsets.ModelViewSet):
    queryset = Container.objects.all()
    serializer_class = ContainerSerializer


class ProductViewSet(viewsets.ModelViewSet):
    queryset = Product.objects.all()
    serializer_class = ProductSerializer


class ProductContainerViewSet(viewsets.ModelViewSet):
    queryset = ProductContainer.objects.all()
    serializer_class = ProductContainerSerializer
