requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
PyYAML>=5.4.1
djangorestframework>=3.12.4
django-filter>=2.4.0
markdown>=3.3.4
